json writing and genre filtering
Added functionality so that when you refresh/close the page, the game object information on the front end is sent to the backend. I also added a field to the front end object called visible that can be used later to tell whether a game should be shown on the webpage. With this, I added a function that sets games to be visible if they match the genres within a list